04_01 진행
missile놔두고 player의 collision부터 해결할것

+++ ImageManager.cpp
@@ -102,24 +102,24 @@ Image* ImageManager::FindImage(const string& key)
 
 Image* ImageManager::AddImage(EImageKey key, const wchar_t* filePath, int width, int height, bool isTransparent, COLORREF transColor)
 {
-    string keyString = std::to_string(key);
+    string keyString = std::to_string((int)key);
     return this->AddImage(keyString, filePath, width, height, isTransparent, transColor);
 }
 
 Image* ImageManager::AddImage(EImageKey key, const wchar_t* filePath, int width, int height, int maxFrameX, int maxFrameY, bool isTransparent, COLORREF transColor)
 {
-    string keyString = std::to_string(key);
+    string keyString = std::to_string((int)key);
     return this->AddImage(keyString, filePath, width, height, maxFrameX, maxFrameY, isTransparent, transColor);
 }
 
 void ImageManager::DeleteImage(EImageKey key)
 {
-    string keyString = std::to_string(key);
+    string keyString = std::to_string((int)key);
     this->DeleteImage(keyString);
 }
 
 Image* ImageManager::FindImage(EImageKey key)
 {
-    string keyString = std::to_string(key);
+    string keyString = std::to_string((int)key);
     return this->FindImage(keyString);
 }
